hardware/rover/requirements.txt
face_recognition>=1.3.0
opencv-python>=4.5.0
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.20.0
Pillow>=8.0.0
requests>=2.25.0
